Updated packages definition to the latest to address security vulnerability
I built a custom component from this template. Github security insight found vulnerabilities (2) with a dependency of one of these package that was used yaml-js. 

I updated package version to their latest and vulnerability was marked as closed. 
I am therefore submitting a pull request to suggest the same operation to be performed here.